Remove references to global APP variable
Instead of relying in a global APP application a function createDataTable
is introduced to create a stand-alone table. The cells will be tagged
with a table identifier (automatically assigned) to avoid conflicts. The
callbacks functions are now local to the function createDataTable and only
refers to veriables local to the function.

Only remaining global variable are essentially dataTableCurrentId and the
function createDataTable itself.

This modification ensures that the implementation is easily usable
in other JS applications without polluting the global namespace.

It does also ensure that multiple tables can exists on the same page.